Home

BandedMatrices.jl Documentation

Creating banded matrices

BandedMatrices.BandedMatrix — Type.

BandedMatrix{T}(uninitialized, (n, m), (l, u))

returns an uninitialized n×m banded matrix of type T with bandwidths (l,u).

source

bones

BandedMatrices.brand — Function.

brand(T,n,m,l,u)

Creates an n×m banded matrix with random numbers in the bandwidth of type T with bandwidths (l,u)

source

To create a banded matrix of all zeros, identity matrix, or with a constant value use the following constructors:

julia> BandedMatrix(Zeros(5,5), (1,2))
5×5 BandedMatrices.BandedMatrix{Float64}:
 0.0  0.0  0.0          
 0.0  0.0  0.0  0.0     
      0.0  0.0  0.0  0.0
           0.0  0.0  0.0
                0.0  0.0

julia> BandedMatrix(Eye(5), (1,2))
5×5 BandedMatrices.BandedMatrix{Float64}:
 1.0  0.0  0.0          
 0.0  1.0  0.0  0.0     
      0.0  1.0  0.0  0.0
           0.0  1.0  0.0
                0.0  1.0

julia> BandedMatrix(Ones(5,5), (1,2))
5×5 BandedMatrices.BandedMatrix{Float64}:
 1.0  1.0  1.0          
 1.0  1.0  1.0  1.0     
      1.0  1.0  1.0  1.0
           1.0  1.0  1.0
                1.0  1.0

Accessing banded matrices

BandedMatrices.bandwidths — Function.

bandwidths(A)

Returns a tuple containing the upper and lower bandwidth of A.

source

BandedMatrices.bandwidth — Function.

bandwidth(A,i)

Returns the lower bandwidth (i==1) or the upper bandwidth (i==2).

source

BandedMatrices.bandrange — Function.

bandrange(A)

Returns the range -bandwidth(A,1):bandwidth(A,2).

source

BandedMatrices.band — Function.

band(i)

Represents the i-th band of a banded matrix.

julia> using BandedMatrices

julia> A = bones(5,5,1,1)
5×5 BandedMatrices.BandedMatrix{Float64}:
 1.0  1.0
 1.0  1.0  1.0
      1.0  1.0  1.0
           1.0  1.0  1.0
                1.0  1.0

julia> A[band(1)]
4-element Array{Float64,1}:
 1.0
 1.0
 1.0
 1.0

source

BandedMatrices.BandRange — Type.

BandRange

Represents the entries in a row/column inside the bands.

julia> using BandedMatrices

julia> A = bones(5,5,1,1)
5×5 BandedMatrices.BandedMatrix{Float64}:
 1.0  1.0
 1.0  1.0  1.0
      1.0  1.0  1.0
           1.0  1.0  1.0
                1.0  1.0

julia> A[2,BandRange]
3-element Array{Float64,1}:
 1.0
 1.0
 1.0

source

Creating symmetric banded matrices

BandedMatrices.SymBandedMatrix — Type.

SymBandedMatrix(T, n, k)

returns an unitialized n×n symmetric banded matrix of type T with bandwidths (-k,k).

source

sbones

BandedMatrices.sbrand — Function.

sbrand(T,n,k)

Creates an n×n symmetric banded matrix with random numbers in the bandwidth of type T with bandwidths (k,k)

source

Banded matrix interface

Banded matrices go beyond the type BandedMatrix: one can also create matrix types that conform to the _banded matrix interface_, in which case many of the utility functions in this package are available. The banded matrix interface consists of the following:

Required methods	Brief description

bandwidth(A, k)	Returns the sub-diagonal bandwidth (k==1) or the super-diagonal bandwidth (k==2)

isbanded(A)	Override to return true

inbands_getindex(A, k, j)	Unsafe: return A[k,j], without the need to check if we are inside the bands

inbands_setindex!(A, v, k, j)	Unsafe: set A[k,j] = v, without the need to check if we are inside the bands

Note that certain SubArrays of BandedMatrix are also banded matrices. The banded matrix interface is implemented for such SubArrays to take advantage of this.

Implementation

Currently, only column-major ordering is supported: a banded matrix B

[ a_11 a_12
  a_21 a_22 a_23
  a_31 a_32 a_33 a_34
       a_42 a_43 a_44  ]

is represented as a BandedMatrix with a field B.data representing the matrix as

[ *     a_12   a_23    a_34
 a_11   a_22   a_33    a_44
 a_21   a_32   a_43    *
 a_31   a_42   *       *       ]

B.l gives the number of subdiagonals (2) and B.u gives the number of super-diagonals (1). Both B.l and B.u must be non-negative at the moment.