BandedMatrices.jl Documentation

Creating banded matrices

BandedMatrices.BandedMatrix — Type.

BandedMatrix(T, n, m, l, u)

returns an unitialized n×m banded matrix of type T with bandwidths (l,u).

source

BandedMatrices.beye — Function.

beye(T,n,l,u)

n×n banded identity matrix of type T with bandwidths (l,u)

source

BandedMatrices.bones — Function.

bones(T,n,m,l,u)

Creates an n×m banded matrix with ones in the bandwidth of type T with bandwidths (l,u)

source

BandedMatrices.brand — Function.

brand(T,n,m,l,u)

Creates an n×m banded matrix with random numbers in the bandwidth of type T with bandwidths (l,u)

source

BandedMatrices.bzeros — Function.

bzeros(T,n,m,l,u)

Creates an n×m banded matrix of all zeros of type T with bandwidths (l,u)

source

Accessing banded matrices

BandedMatrices.bandwidths — Function.

bandwidths(A)

Returns a tuple containing the upper and lower bandwidth of A.

source

BandedMatrices.bandwidth — Function.

bandwidth(A,i)

Returns the lower bandwidth (i==1) or the upper bandwidth (i==2).

source

BandedMatrices.bandrange — Function.

bandrange(A)

Returns the range -bandwidth(A,1):bandwidth(A,2).

source

BandedMatrices.band — Function.

band(i)

Represents the i-th band of a banded matrix.

julia> using BandedMatrices

julia> A = bones(5,5,1,1)
5×5 BandedMatrices.BandedMatrix{Float64}:
 1.0  1.0
 1.0  1.0  1.0
      1.0  1.0  1.0
           1.0  1.0  1.0
                1.0  1.0

julia> A[band(1)]
4-element Array{Float64,1}:
 1.0
 1.0
 1.0
 1.0

source

BandedMatrices.BandRange — Type.

BandRange

Represents the entries in a row/column inside the bands.

julia> using BandedMatrices

julia> A = bones(5,5,1,1)
5×5 BandedMatrices.BandedMatrix{Float64}:
 1.0  1.0
 1.0  1.0  1.0
      1.0  1.0  1.0
           1.0  1.0  1.0
                1.0  1.0

julia> A[2,BandRange]
3-element Array{Float64,1}:
 1.0
 1.0
 1.0

source